Math: move IEquation and IDisplay to shared Abstractions
- Add Fluxion.Math/Abstractions/EquationAbstractions.cs with IEquation + IDisplay
- Remove legacy interface copies from Algebra/Trigonometry
- Update Algebra/Trigonometry equation wrappers to use Fluxion.Math.Abstractions
- (Optional) Add GlobalUsings.cs for simpler imports
