+++ Fluxion.Math/Algebra/Equations/Polynomial.cs
@@ -2,7 +2,7 @@
 using System;
 using System.Linq;
 using Fluxion.Math.Algebra.Concepts;
-
+using Fluxion.Math.Abstractions;
 namespace Fluxion.Math.Algebra.Equations
 {
     /// <summary>
