+++ Fluxion.Math/Algebra/Equations/SystemOfEquations2x2.cs
@@ -1,6 +1,6 @@
 ﻿// Fluxion.Math/Algebra/Equations/SystemOfEquations2x2.cs
 using Fluxion.Math.Algebra.Concepts;
-
+using Fluxion.Math.Abstractions;
 namespace Fluxion.Math.Algebra.Equations
 {
     public sealed class SystemOfEquations2x2 : IDisplay
