+++ Fluxion.Math/Algebra/Operations/Substitution.cs
@@ -1,6 +1,6 @@
 ﻿// Fluxion.Math/Algebra/Operations/Substitution.cs
 using Fluxion.Math.Algebra.Concepts;
-
+using Fluxion.Math.Abstractions;
 namespace Fluxion.Math.Algebra.Operations
 {
     public static class Substitution
