+++ Fluxion.Math/Trigonometry/Functions/CosineEquation.cs
@@ -1,10 +1,10 @@
-﻿//Fluxion.Math/Trigonometry/Functions/CosineEquation.cs
-using Fluxion.Math.Algebra.Concepts;
+﻿using Fluxion.Math.Abstractions;
 using Fluxion.Math.Trigonometry.Concepts;
 using System;
 
 namespace Fluxion.Math.Trigonometry.Functions
 {
+    /// <summary>f(x) = Amplitude * cos(Frequency * x + Phase) + VerticalOffset</summary>
     public readonly struct CosineEquation : IEquation, IDisplay
     {
         public CosineModel Model { get; }
@@ -15,8 +15,8 @@ namespace Fluxion.Math.Trigonometry.Functions
             => Model.Amplitude * System.Math.Cos(Model.Frequency * x + Model.Phase) + Model.VerticalOffset;
 
         public bool IsWellFormed()
-            => !(double.IsNaN(Model.Amplitude) || double.IsNaN(Model.Frequency)
-                 || double.IsNaN(Model.Phase) || double.IsNaN(Model.VerticalOffset));
+            => !(double.IsNaN(Model.Amplitude) || double.IsNaN(Model.Frequency) ||
+                 double.IsNaN(Model.Phase) || double.IsNaN(Model.VerticalOffset));
 
         public string AsString(string variable = "x")
             => $"{Model.Amplitude}·cos({Model.Frequency}·{variable} + {Model.Phase}) + {Model.VerticalOffset}";
